Move automapper cache to separate class
